move GeminiKernelFunctionMetadataExtensions to Google connector namespace
