+++ dotnet/src/Connectors/Connectors.Google.UnitTests/GeminiToolCallBehaviorTests.cs
@@ -6,6 +6,7 @@ using System.Text.Json;
 using Microsoft.SemanticKernel;
 using Microsoft.SemanticKernel.Connectors.Google;
 using Microsoft.SemanticKernel.Connectors.Google.Core;
+using Microsoft.SemanticKernel.Connectors.Google.Extensions;
 using Xunit;
 
 namespace SemanticKernel.Connectors.Google.UnitTests;
